+++ Dockerfile
@@ -5,7 +5,7 @@ RUN apt-get update
 RUN curl -sL https://deb.nodesource.com/setup_10.x | bash -
 RUN apt-get update && apt-get install -y nodejs
 
-WORKDIR /aws-amazon-ecs-cli-v2
+WORKDIR /copilot
 COPY . .
 RUN go env -w GOPROXY=direct
 RUN make release
