+++ e2e/root/root_test.go
@@ -24,7 +24,7 @@ var _ = Describe("Root", func() {
 			Expect(err).NotTo(HaveOccurred())
 			// Versions look like ecs-preview version: v0.0.4-34-g133b977
 			// the extra bit at the end is if the build isn't a tagged release.
-			Expect(output).To(MatchRegexp(`ecs-preview version: v\d*\.\d*\.\d*.*`))
+			Expect(output).To(MatchRegexp(`copilot version: v\d*\.\d*\.\d*.*`))
 		})
 	})
 })
